refactor: remove migrations
