+++ apps/server/src/api-data/excel/excel.router.ts
@@ -12,5 +12,3 @@ export const router = express.Router();
 router.post('/upload', uploadExcel, validateFileExists, postExcel);
 router.get('/worksheets', getWorksheets);
 router.post('/preview', validateImportMapOptions, previewExcel);
-
-// TODO: validate import map
