+++ apps/server/src/api-data/rundown/rundown.controller.ts
@@ -1,11 +1,4 @@
-import {
-  ErrorResponse,
-  MessageResponse,
-  OntimeRundown,
-  OntimeRundownEntry,
-  RundownCached,
-  RundownPaginated,
-} from 'ontime-types';
+import { ErrorResponse, MessageResponse, OntimeRundown, OntimeRundownEntry, RundownCached } from 'ontime-types';
 import { getErrorMessage } from 'ontime-utils';
 
 import type { Request, Response } from 'express';
@@ -21,12 +14,7 @@ import {
   reorderEvent,
   swapEvents,
 } from '../../services/rundown-service/RundownService.js';
-import {
-  getEventWithId,
-  getNormalisedRundown,
-  getPaginated,
-  getRundown,
-} from '../../services/rundown-service/rundownUtils.js';
+import { getEventWithId, getNormalisedRundown, getRundown } from '../../services/rundown-service/rundownUtils.js';
 
 export async function rundownGetAll(_req: Request, res: Response<OntimeRundown>) {
   const rundown = getRundown();
@@ -52,34 +40,6 @@ export async function rundownGetById(req: Request, res: Response<OntimeRundownEn
   } catch (error) {
     const message = getErrorMessage(error);
     res.status(500).json({ message });
-  }
-}
-
-export async function rundownGetPaginated(req: Request, res: Response<RundownPaginated | ErrorResponse>) {
-  const { limit, offset } = req.query;
-
-  if (limit == null && offset == null) {
-    return res.json({
-      rundown: getRundown(),
-      total: getRundown().length,
-    });
-  }
-
-  try {
-    let parsedOffset = Number(offset);
-    if (Number.isNaN(parsedOffset)) {
-      parsedOffset = 0;
-    }
-    let parsedLimit = Number(limit);
-    if (Number.isNaN(parsedLimit)) {
-      parsedLimit = Infinity;
-    }
-    const paginatedRundown = getPaginated(parsedOffset, parsedLimit);
-
-    res.status(200).json(paginatedRundown);
-  } catch (error) {
-    const message = getErrorMessage(error);
-    res.status(400).json({ message });
   }
 }
