+++ apps/server/src/api-data/rundown/rundown.router.ts
@@ -8,7 +8,6 @@ import {
   rundownGetAll,
   rundownGetById,
   rundownGetNormalised,
-  rundownGetPaginated,
   rundownPost,
   rundownPut,
   rundownReorder,
@@ -18,7 +17,6 @@ import {
   paramsMustHaveEventId,
   rundownArrayOfIds,
   rundownBatchPutValidator,
-  rundownGetPaginatedQueryParams,
   rundownPostValidator,
   rundownPutValidator,
   rundownReorderValidator,
@@ -28,7 +26,6 @@ import {
 export const router = express.Router();
 
 router.get('/', rundownGetAll); // not used in Ontime frontend
-router.get('/paginated', rundownGetPaginatedQueryParams, rundownGetPaginated); // not used in Ontime frontend
 router.get('/normalised', rundownGetNormalised);
 router.get('/:eventId', paramsMustHaveEventId, rundownGetById); // not used in Ontime frontend
