+++ apps/server/src/api-data/rundown/rundown.validation.ts
@@ -1,4 +1,4 @@
-import { body, param, query, validationResult } from 'express-validator';
+import { body, param, validationResult } from 'express-validator';
 import type { Request, Response, NextFunction } from 'express';
 
 export const rundownPostValidator = [
@@ -70,17 +70,6 @@ export const paramsMustHaveEventId = [
 export const rundownArrayOfIds = [
   body('ids').isArray().exists(),
   body('ids.*').isString(),
-
-  (req: Request, res: Response, next: NextFunction) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) return res.status(422).json({ errors: errors.array() });
-    next();
-  },
-];
-
-export const rundownGetPaginatedQueryParams = [
-  query('offset').isNumeric().optional(),
-  query('limit').isNumeric().optional(),
 
   (req: Request, res: Response, next: NextFunction) => {
     const errors = validationResult(req);
