+++ packages/types/src/api/ontime-controller/BackendResponse.type.ts
@@ -1,4 +1,3 @@
-import type { OntimeRundown } from '../../definitions/core/Rundown.type.js';
 import type { Playback } from '../../definitions/runtime/Playback.type.js';
 import type { MaybeString } from '../../utils/utils.type.js';
 
@@ -51,8 +50,3 @@ export type ProjectLogoResponse = {
 export type ErrorResponse = MessageResponse;
 
 export type AuthenticationStatus = 'authenticated' | 'not_authenticated' | 'pending';
-
-export type RundownPaginated = {
-  rundown: OntimeRundown;
-  total: number;
-};
