+++ packages/types/src/index.ts
@@ -67,7 +67,6 @@ export type {
   ErrorResponse,
   ProjectFileListResponse,
   MessageResponse,
-  RundownPaginated,
   SessionStats,
   ProjectLogoResponse,
 } from './api/ontime-controller/BackendResponse.type.js';
